+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - application name is now optional: when unset, multiple instance can run in parallel.
 - the `--network` argument now takes an argument: private, host or a shared name.
 - add `--ps` argument to list running app.
+- add support for nix flake binary cache.
 
 ## 0.2.0 (2022-07-26)